Rename file correctly
